Refactor of logger
