--- src/main/java/dk/cngroup/lentils/logger/Printer.java
@@ -1,5 +1,0 @@
-package dk.cngroup.lentils.logger;
-
-public interface Printer {
-    void println(String message);
-}
