--- src/main/java/dk/cngroup/lentils/logger/StdoutPrinter.java
@@ -1,11 +1,0 @@
-package dk.cngroup.lentils.logger;
-
-import org.springframework.stereotype.Component;
-
-@Component
-public class StdoutPrinter implements Printer {
-    @Override
-    public void println(String message) {
-        System.out.println(message);
-    }
-}
