+++ src/main/java/dk/cngroup/lentils/logger/printer/Printer.java
@@ -1,0 +1,5 @@
+package dk.cngroup.lentils.logger.printer;
+
+public interface Printer {
+    <T> void println(T message);
+}
